backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL. On Windows, ensure PostgreSQL is installed.

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Config and validation
pydantic-settings==2.1.0
email-validator==2.1.0.post1
